fix : 농장 기상청 데이터 API 이동
